refacto: replace add contactEmails form section by a preffilled text field
