+++ client/src/lib/components/DatasetForm/DatasetForm.spec.ts
@@ -4,10 +4,8 @@ import DatasetForm from "./DatasetForm.svelte";
 import {
   render,
   fireEvent,
-  getByRole as getByRoleIn,
-  getByLabelText,
 } from "@testing-library/svelte";
-import type { DataFormat, DatasetFormData } from "src/definitions/datasets";
+import type { DataFormat } from "src/definitions/datasets";
 import { UPDATE_FREQUENCY } from "src/constants";
 import { getFakeDataSetFormData } from "src/fixtures/dataset";
 
@@ -60,35 +58,6 @@ describe("Test the dataset form", () => {
     expect(inputs.length).toBe(1);
     expect(inputs[0]).not.toBeRequired();
     expect(inputs[0]).toHaveAttribute("type", "email");
-  });
-
-  test("Contact emails can be added, filled and removed", async () => {
-    const { getAllByLabelText, getByRole } = render(DatasetForm);
-    let inputs = getAllByLabelText(/Contact \d/);
-    expect(inputs.length).toBe(1);
-    expect(inputs[0]).toHaveValue("");
-
-    const addButton = getByRole("button", { name: /Ajouter/i });
-    await fireEvent.click(addButton);
-    inputs = getAllByLabelText(/Contact \d/);
-    expect(inputs.length).toBe(2);
-    expect(inputs[0]).toHaveValue("");
-    expect(inputs[1]).toHaveValue("");
-
-    await fireEvent.input(inputs[1], {
-      target: { value: "contact@example.org" },
-    });
-    inputs = getAllByLabelText(/Contact \d/);
-    expect(inputs[0]).toHaveValue("");
-    expect(inputs[1]).toHaveValue("contact@example.org");
-
-    const removeButton = getByRoleIn(inputs[1].parentElement, "button", {
-      name: /Supprimer/i,
-    });
-    await fireEvent.click(removeButton);
-    inputs = getAllByLabelText(/Contact \d/);
-    expect(inputs.length).toBe(1);
-    expect(inputs[0]).toHaveValue("");
   });
 
   test('The "service field" field is present', () => {
